requirements.txt
pandas
scikit-learn
PyYAML
joblib
pytest
biopython
umap-learn
matplotlib
seaborn